# pkg/trace/obfuscate: optimize the SQL obfuscator