--- a/pkg/trace/obfuscate/sql.go
+++ b/pkg/trace/obfuscate/sql.go
@@ -266,14 +266,14 @@ func (oq *ObfuscatedQuery) Cost() int64 {
 // given set of filters.
 func attemptObfuscation(tokenizer *SQLTokenizer) (*ObfuscatedQuery, error) {
 	var (
-		tableFinder       = &tableFinderFilter{}
-		filterTableFinder = config.HasFeature("table_names")
-		out               = *bytes.NewBuffer(make([]byte, 0, len(tokenizer.buf)))
-		err               error
-		lastToken         TokenKind
-		discard           discardFilter
-		replace           replaceFilter
-		grouping          groupingFilter
+		tableFinder    = &tableFinderFilter{}
+		useTableFinder = config.HasFeature("table_names")
+		out            = *bytes.NewBuffer(make([]byte, 0, len(tokenizer.buf)))
+		err            error
+		lastToken      TokenKind
+		discard        discardFilter
+		replace        replaceFilter
+		grouping       groupingFilter
 	)
 	// call Scan() function until tokens are available or if a LEX_ERROR is raised. After
 	// retrieving a token, send it to the tokenFilter chains so that the token is discarded
@@ -296,7 +296,7 @@ func attemptObfuscation(tokenizer *SQLTokenizer) (*ObfuscatedQuery, error) {
 		if token, buff, err = grouping.Filter(token, lastToken, buff); err != nil {
 			return nil, err
 		}
-		if filterTableFinder {
+		if useTableFinder {
 			if token, buff, err = tableFinder.Filter(token, lastToken, buff); err != nil {
 				return nil, err
 			}