--- a/pkg/trace/obfuscate/sql_tokenizer.go
+++ b/pkg/trace/obfuscate/sql_tokenizer.go
@@ -8,7 +8,6 @@ package obfuscate
 import (
 	"bytes"
 	"fmt"
-	"strings"
 	"unicode"
 	"unicode/utf8"
 )
@@ -80,10 +79,11 @@ const escapeCharacter = '\\'
 // SQLTokenizer is the struct used to generate SQL
 // tokens for the parser.
 type SQLTokenizer struct {
-	rd       *strings.Reader // the "rune" reader
-	pos      int             // byte offset of lastChar
-	lastChar rune            // last read rune
-	err      error           // any error occurred while reading
+	pos      int    // byte offset of lastChar
+	lastChar rune   // last read rune
+	buf      []byte // buf holds the query that we are parsing
+	index    int    // index is the index into buf where the unread portion of the query begins.
+	err      error  // any error occurred while reading
 
 	literalEscapes bool // indicates we should not treat backslashes as escape characters
 	seenEscape     bool // indicates whether this tokenizer has seen an escape character within a string
@@ -93,16 +93,17 @@ type SQLTokenizer struct {
 // whether escape characters should be treated literally or as such.
 func NewSQLTokenizer(sql string, literalEscapes bool) *SQLTokenizer {
 	return &SQLTokenizer{
-		rd:             strings.NewReader(sql),
+		buf:            []byte(sql),
 		literalEscapes: literalEscapes,
 	}
 }
 
 // Reset the underlying buffer and positions
 func (tkn *SQLTokenizer) Reset(in string) {
-	tkn.rd.Reset(in)
 	tkn.pos = 0
 	tkn.lastChar = 0
+	tkn.buf = []byte(in)
+	tkn.index = 0
 	tkn.err = nil
 }
 
@@ -159,12 +160,12 @@ func (tkn *SQLTokenizer) Scan() (TokenKind, []byte) {
 			}
 			fallthrough
 		case '=', ',', ';', '(', ')', '+', '*', '&', '|', '^', '~', '[', ']', '?':
-			return TokenKind(ch), runeBytes(ch)
+			return TokenKind(ch), tkn.grab()
 		case '.':
 			if isDigit(tkn.lastChar) {
 				return tkn.scanNumber(true)
 			}
-			return TokenKind(ch), runeBytes(ch)
+			return TokenKind(ch), tkn.grab()
 		case '/':
 			switch tkn.lastChar {
 			case '/':
@@ -174,14 +175,14 @@ func (tkn *SQLTokenizer) Scan() (TokenKind, []byte) {
 				tkn.next()
 				return tkn.scanCommentType2()
 			default:
-				return TokenKind(ch), runeBytes(ch)
+				return TokenKind(ch), tkn.grab()
 			}
 		case '-':
 			if tkn.lastChar == '-' {
 				tkn.next()
 				return tkn.scanCommentType1("--")
 			}
-			return TokenKind(ch), runeBytes(ch)
+			return TokenKind(ch), tkn.grab()
 		case '#':
 			tkn.next()
 			return tkn.scanCommentType1("#")
@@ -200,21 +201,21 @@ func (tkn *SQLTokenizer) Scan() (TokenKind, []byte) {
 					return LE, []byte("<=")
 				}
 			default:
-				return TokenKind(ch), runeBytes(ch)
+				return TokenKind(ch), tkn.grab()
 			}
 		case '>':
 			if tkn.lastChar == '=' {
 				tkn.next()
 				return GE, []byte(">=")
 			}
-			return TokenKind(ch), runeBytes(ch)
+			return TokenKind(ch), tkn.grab()
 		case '!':
 			if tkn.lastChar == '=' {
 				tkn.next()
 				return NE, []byte("!=")
 			}
 			tkn.setErr(`expected "=" after "!", got "%c" (%d)`, tkn.lastChar, tkn.lastChar)
-			return LexError, []byte("!")
+			return LexError, tkn.grab()
 		case '\'':
 			return tkn.scanString(ch, String)
 		case '"':
@@ -230,14 +231,14 @@ func (tkn *SQLTokenizer) Scan() (TokenKind, []byte) {
 				return tkn.scanFormatParameter('%')
 			}
 			// modulo operator (e.g. 'id % 8')
-			return TokenKind(ch), runeBytes(ch)
+			return TokenKind(ch), tkn.grab()
 		case '$':
 			return tkn.scanPreparedStatement('$')
 		case '{':
 			return tkn.scanEscapeSequence('{')
 		default:
 			tkn.setErr(`unexpected byte %d`, ch)
-			return LexError, runeBytes(ch)
+			return LexError, tkn.grab()
 		}
 	}
 }
@@ -246,99 +247,116 @@ func (tkn *SQLTokenizer) skipBlank() {
 	for unicode.IsSpace(tkn.lastChar) {
 		tkn.next()
 	}
+	tkn.grab()
+}
+
+// myFastToUpper returns a copy of the byte slice s with all Unicode letters mapped to
+// their upper case. It uses the array pointed to by dst as the destination if possible.
+func myFastToUpper(src, dst []byte) []byte {
+	dst = dst[:0]
+	isASCII, hasLower := true, false
+	for i := 0; i < len(src); i++ {
+		c := src[i]
+		if c >= utf8.RuneSelf {
+			isASCII = false
+			break
+		}
+		hasLower = hasLower || ('a' <= c && c <= 'z')
+	}
+	if cap(dst) < len(src) {
+		dst = make([]byte, 0, len(src))
+	}
+	if isASCII { // optimize for ASCII-only byte slices.
+		if !hasLower {
+			// Just return a copy.
+			return append(dst, src...)
+		}
+		dst = dst[:len(src)]
+		//b := make([]byte, len(src))
+		for i := 0; i < len(src); i++ {
+			c := src[i]
+			if 'a' <= c && c <= 'z' {
+				c -= 'a' - 'A'
+			}
+			dst[i] = c
+		}
+		return dst
+	}
+	// This *could* be optimized, but it's an uncommon case.
+	return bytes.Map(unicode.ToUpper, src)
 }
 
 func (tkn *SQLTokenizer) scanIdentifier() (TokenKind, []byte) {
-	buffer := &bytes.Buffer{}
-	buffer.WriteRune(tkn.lastChar)
 	tkn.next()
-
 	for isLetter(tkn.lastChar) || isDigit(tkn.lastChar) || tkn.lastChar == '.' || tkn.lastChar == '*' {
-		buffer.WriteRune(tkn.lastChar)
 		tkn.next()
 	}
-	upper := bytes.ToUpper(buffer.Bytes())
+
+	t := tkn.grab()
+	// Space allows us to upper-case identifiers 200 bytes long or less without allocating.
+	var space [200]byte
+	upper := myFastToUpper(t, space[:0])
 	if keywordID, found := keywords[string(upper)]; found {
-		return keywordID, buffer.Bytes()
+		return keywordID, t
 	}
-	return ID, buffer.Bytes()
+	return ID, t
 }
 
 func (tkn *SQLTokenizer) scanLiteralIdentifier(quote rune) (TokenKind, []byte) {
-	buffer := &bytes.Buffer{}
-	buffer.WriteRune(tkn.lastChar)
+	tkn.grab() // throw away initial quote
 	if !isLetter(tkn.lastChar) && !isDigit(tkn.lastChar) {
 		tkn.setErr(`unexpected character "%c" (%d) in literal identifier`, tkn.lastChar, tkn.lastChar)
-		return LexError, buffer.Bytes()
+		return LexError, tkn.grab()
 	}
 	for tkn.next(); skipNonLiteralIdentifier(tkn.lastChar); tkn.next() {
-		buffer.WriteRune(tkn.lastChar)
 	}
+
+	t := tkn.grab()
 	// literals identifier are enclosed between quotes
 	if tkn.lastChar != quote {
 		tkn.setErr(`literal identifiers must end in "%c", got "%c" (%d)`, quote, tkn.lastChar, tkn.lastChar)
-		return LexError, buffer.Bytes()
+		return LexError, t
 	}
 	tkn.next()
-	return ID, buffer.Bytes()
+	return ID, t
 }
 
 func (tkn *SQLTokenizer) scanVariableIdentifier(prefix rune) (TokenKind, []byte) {
-	buffer := &bytes.Buffer{}
-	buffer.WriteRune(prefix)
-	buffer.WriteRune(tkn.lastChar)
-
 	for tkn.next(); tkn.lastChar != ')' && tkn.lastChar != EOFChar; tkn.next() {
-		buffer.WriteRune(tkn.lastChar)
 	}
-
-	buffer.WriteRune(tkn.lastChar)
 	tkn.next()
-	buffer.WriteRune(tkn.lastChar)
 	if !isLetter(tkn.lastChar) {
 		tkn.setErr(`invalid character after variable identifier: "%c" (%d)`, tkn.lastChar, tkn.lastChar)
-		return LexError, buffer.Bytes()
+		return LexError, tkn.grab()
 	}
 	tkn.next()
-	return Variable, buffer.Bytes()
+	return Variable, tkn.grab()
 }
 
 func (tkn *SQLTokenizer) scanFormatParameter(prefix rune) (TokenKind, []byte) {
-	buffer := &bytes.Buffer{}
-	buffer.WriteRune(prefix)
-	buffer.WriteRune(tkn.lastChar)
-
 	tkn.next()
-	return Variable, buffer.Bytes()
+	return Variable, tkn.grab()
 }
 
 func (tkn *SQLTokenizer) scanPreparedStatement(prefix rune) (TokenKind, []byte) {
-	buffer := &bytes.Buffer{}
-
 	// a prepared statement expect a digit identifier like $1
 	if !isDigit(tkn.lastChar) {
 		tkn.setErr(`prepared statements must start with digits, got "%c" (%d)`, tkn.lastChar, tkn.lastChar)
-		return LexError, buffer.Bytes()
+		return LexError, tkn.grab()
 	}
 
+	// scanNumber keeps the prefix rune intact.
 	// read numbers and return an error if any
 	token, buff := tkn.scanNumber(false)
 	if token == LexError {
 		tkn.setErr("invalid number")
-		return LexError, buffer.Bytes()
+		return LexError, tkn.grab()
 	}
-
-	buffer.WriteRune(prefix)
-	buffer.Write(buff)
-	return PreparedStatement, buffer.Bytes()
+	return PreparedStatement, buff
 }
 
 func (tkn *SQLTokenizer) scanEscapeSequence(braces rune) (TokenKind, []byte) {
-	buffer := &bytes.Buffer{}
-	buffer.WriteRune(braces)
-
 	for tkn.lastChar != '}' && tkn.lastChar != EOFChar {
-		buffer.WriteRune(tkn.lastChar)
 		tkn.next()
 	}
 
@@ -346,62 +364,56 @@ func (tkn *SQLTokenizer) scanEscapeSequence(braces rune) (TokenKind, []byte) {
 	// the closing curly braces
 	if tkn.lastChar == EOFChar {
 		tkn.setErr("unexpected EOF in escape sequence")
-		return LexError, buffer.Bytes()
+		return LexError, tkn.grab()
 	}
 
-	buffer.WriteRune(tkn.lastChar)
 	tkn.next()
-	return EscapeSequence, buffer.Bytes()
+	return EscapeSequence, tkn.grab()
 }
 
 func (tkn *SQLTokenizer) scanBindVar() (TokenKind, []byte) {
-	buffer := bytes.NewBufferString(":")
 	token := ValueArg
 	if tkn.lastChar == ':' {
 		token = ListArg
-		buffer.WriteRune(tkn.lastChar)
 		tkn.next()
 	}
 	if !isLetter(tkn.lastChar) {
 		tkn.setErr(`bind variables should start with letters, got "%c" (%d)`, tkn.lastChar, tkn.lastChar)
-		return LexError, buffer.Bytes()
+		return LexError, tkn.grab()
 	}
 	for isLetter(tkn.lastChar) || isDigit(tkn.lastChar) || tkn.lastChar == '.' {
-		buffer.WriteRune(tkn.lastChar)
 		tkn.next()
 	}
-	return token, buffer.Bytes()
+	return token, tkn.grab()
 }
 
-func (tkn *SQLTokenizer) scanMantissa(base int, buffer *bytes.Buffer) {
+func (tkn *SQLTokenizer) scanMantissa(base int) {
 	for digitVal(tkn.lastChar) < base {
-		tkn.consumeNext(buffer)
+		tkn.next()
 	}
 }
 
 func (tkn *SQLTokenizer) scanNumber(seenDecimalPoint bool) (TokenKind, []byte) {
-	buffer := &bytes.Buffer{}
 	if seenDecimalPoint {
-		buffer.WriteByte('.')
-		tkn.scanMantissa(10, buffer)
+		tkn.scanMantissa(10)
 		goto exponent
 	}
 
 	if tkn.lastChar == '0' {
 		// int or float
-		tkn.consumeNext(buffer)
+		tkn.next()
 		if tkn.lastChar == 'x' || tkn.lastChar == 'X' {
 			// hexadecimal int
-			tkn.consumeNext(buffer)
-			tkn.scanMantissa(16, buffer)
+			tkn.next()
+			tkn.scanMantissa(16)
 		} else {
 			// octal int or float
 			seenDecimalDigit := false
-			tkn.scanMantissa(8, buffer)
+			tkn.scanMantissa(8)
 			if tkn.lastChar == '8' || tkn.lastChar == '9' {
 				// illegal octal int or float
 				seenDecimalDigit = true
-				tkn.scanMantissa(10, buffer)
+				tkn.scanMantissa(10)
 			}
 			if tkn.lastChar == '.' || tkn.lastChar == 'e' || tkn.lastChar == 'E' {
 				goto fraction
@@ -409,35 +421,36 @@ func (tkn *SQLTokenizer) scanNumber(seenDecimalPoint bool) (TokenKind, []byte) {
 			// octal int
 			if seenDecimalDigit {
 				// tkn.setErr called in caller
-				return LexError, buffer.Bytes()
+				return LexError, tkn.grab()
 			}
 		}
 		goto exit
 	}
 
 	// decimal int or float
-	tkn.scanMantissa(10, buffer)
+	tkn.scanMantissa(10)
 
 fraction:
 	if tkn.lastChar == '.' {
-		tkn.consumeNext(buffer)
-		tkn.scanMantissa(10, buffer)
+		tkn.next()
+		tkn.scanMantissa(10)
 	}
 
 exponent:
 	if tkn.lastChar == 'e' || tkn.lastChar == 'E' {
-		tkn.consumeNext(buffer)
+		tkn.next()
 		if tkn.lastChar == '+' || tkn.lastChar == '-' {
-			tkn.consumeNext(buffer)
+			tkn.next()
 		}
-		tkn.scanMantissa(10, buffer)
+		tkn.scanMantissa(10)
 	}
 
 exit:
-	if buffer.Len() == 0 {
+	t := tkn.grab()
+	if len(t) == 0 {
 		return LexError, nil
 	}
-	return Number, buffer.Bytes()
+	return Number, t
 }
 
 func (tkn *SQLTokenizer) scanString(delim rune, kind TokenKind) (TokenKind, []byte) {
@@ -480,60 +493,68 @@ func (tkn *SQLTokenizer) scanString(delim rune, kind TokenKind) (TokenKind, []by
 }
 
 func (tkn *SQLTokenizer) scanCommentType1(prefix string) (TokenKind, []byte) {
-	buffer := &bytes.Buffer{}
-	buffer.WriteString(prefix)
 	for tkn.lastChar != EOFChar {
 		if tkn.lastChar == '\n' {
-			tkn.consumeNext(buffer)
+			tkn.next()
 			break
 		}
-		tkn.consumeNext(buffer)
+		tkn.next()
 	}
-	return Comment, buffer.Bytes()
+	return Comment, tkn.grab()
 }
 
 func (tkn *SQLTokenizer) scanCommentType2() (TokenKind, []byte) {
-	buffer := &bytes.Buffer{}
-	buffer.WriteString("/*")
 	for {
 		if tkn.lastChar == '*' {
-			tkn.consumeNext(buffer)
+			tkn.next()
 			if tkn.lastChar == '/' {
-				tkn.consumeNext(buffer)
+				tkn.next()
 				break
 			}
 			continue
 		}
 		if tkn.lastChar == EOFChar {
 			tkn.setErr("unexpected EOF in comment")
-			return LexError, buffer.Bytes()
+			return LexError, tkn.grab()
 		}
-		tkn.consumeNext(buffer)
+		tkn.next()
 	}
-	return Comment, buffer.Bytes()
+	return Comment, tkn.grab()
 }
 
-func (tkn *SQLTokenizer) consumeNext(buffer *bytes.Buffer) {
-	if tkn.lastChar == EOFChar {
-		// This should never happen.
-		panic("unexpected EOF")
-	}
-	buffer.WriteRune(tkn.lastChar)
-	tkn.next()
-}
-
+// Next sets tkn.lastChar to the next rune at tkn.buf[tkn.index], and advances the tkn.pos and
+// tkn.index by its size. If next() is unable to scan a character, tkn.lastChar is set to EOFChar.
 func (tkn *SQLTokenizer) next() {
-	ch, _, err := tkn.rd.ReadRune()
+	ch, n := utf8.DecodeRune(tkn.buf[tkn.index:])
+	if ch == utf8.RuneError && n == 0 {
+		// only EOF is possible
+		tkn.pos++
+		tkn.lastChar = EOFChar
+		return
+	}
+
+	len := utf8.RuneLen(ch)
 	if tkn.lastChar != 0 || tkn.pos > 0 {
 		// we are past the first character
-		tkn.pos += utf8.RuneLen(ch)
+		tkn.pos += len
 	}
-	if err != nil {
-		// only EOF is possible
-		tkn.lastChar = EOFChar
-	} else {
-		tkn.lastChar = ch
+	tkn.index += len
+	tkn.lastChar = ch
+}
+
+// grab returns a byte slice of all the characters advanced over with next() since the last call to
+// grab.
+func (tkn *SQLTokenizer) grab() []byte {
+	if tkn.lastChar == EOFChar {
+		ret := tkn.buf[:tkn.index]
+		tkn.buf = tkn.buf[tkn.index:]
+		tkn.index = 0
+		return ret
 	}
+	ret := tkn.buf[:tkn.index-1]
+	tkn.buf = tkn.buf[tkn.index-1:]
+	tkn.index = 1
+	return ret
 }
 
 func skipNonLiteralIdentifier(ch rune) bool {