--- a/releasenotes/notes/refactor-sql-obfuscator-3c8a0d1ef428c410.yaml
+++ b/releasenotes/notes/refactor-sql-obfuscator-3c8a0d1ef428c410.yaml
@@ -1,0 +1,11 @@
+# Each section from every releasenote are combined when the
+# CHANGELOG.rst is rendered. So the text needs to be worded so that
+# it does not depend on any information only available in another
+# section. This may mean repeating some details, but each section
+# must be readable independently of the other.
+#
+# Each section note must be formatted as reStructuredText.
+---
+enhancements:
+  - |
+    APM: refactor the SQL obfuscator to be significantly more efficient.