--- a/releasenotes/notes/refactor-sql-obfuscator-3c8a0d1ef428c410.yaml
+++ b/releasenotes/notes/refactor-sql-obfuscator-3c8a0d1ef428c410.yaml
@@ -8,4 +8,4 @@
 ---
 enhancements:
   - |
-    APM: refactor the SQL obfuscator to be significantly more efficient.
+    APM: refactored the SQL obfuscator to be significantly more efficient.