### What does this PR do?

Previously, the SQL obfuscator would build a new byte array with
bytes.Buffer for every token. This resulted in a lot of allocations. The
new obfuscator returns slices into the original query rather than new
allocations, saving a lot of work. Comparing to the previous obfuscator,
the new one is significantly better in both space and time.
```
% benchcmp master.bench new.bench
benchmark                                      old ns/op     new ns/op     delta
BenchmarkObfuscateSQLString/query1/3694-8     119274        76567         -35.81%
BenchmarkObfuscateSQLString/query3/980-8      34102         20726         -39.22%
BenchmarkObfuscateSQLString/Escaping/512-8     12644         8888          -29.71%
BenchmarkObfuscateSQLString/Grouping/199-8     6353          4858          -23.53%
BenchmarkObfuscateSQLString/random-8           1929          1378          -28.56%

benchmark                                      old allocs     new allocs     delta
BenchmarkObfuscateSQLString/query1/3694-8     912            83             -90.90%
BenchmarkObfuscateSQLString/query3/980-8      242            6              -97.52%
BenchmarkObfuscateSQLString/Escaping/512-8     93             12             -87.10%
BenchmarkObfuscateSQLString/Grouping/199-8     74             10             -86.49%
BenchmarkObfuscateSQLString/random-8           24             8              -66.67%

benchmark                                      old bytes     new bytes     delta
BenchmarkObfuscateSQLString/query1/3694-8     43489         16256         -62.62%
BenchmarkObfuscateSQLString/query3/980-8      11264         3008          -73.30%
BenchmarkObfuscateSQLString/Escaping/512-8     3264          1936          -40.69%
BenchmarkObfuscateSQLString/Grouping/199-8     1920          816           -57.50%
BenchmarkObfuscateSQLString/random-8           752           248           -67.02%
```

### Motivation

The SQL obfuscator has been the source of frequent performance issues. 

### Additional Notes

Anything else we should know when reviewing?

### Describe your test plan

The test suite passes, but since this is a large change we probably want to do some testing against real applications.