@clutchski 2020-09-30
Obviously testing is important as you mentioned. Great work. I like the attention to low level details.