@knusbaum 2020-10-15
Cleaned up and added another test for our new `toUpper`